refactor: 모듈 분리
#1 #2
