+++ config.py
@@ -1,0 +1,6 @@
+import os
+from dotenv import load_dotenv
+
+load_dotenv()
+
+KIOSK_APP_URL = os.getenv("KIOSK_APP_URL")
