+++ interfaces/output.py
@@ -1,0 +1,6 @@
+from typing import Optional
+from pydantic import BaseModel, Field
+
+class Output(BaseModel):
+    status_code: Optional[int] = Field(default=None)
+    message: Optional[str] = Field(default=None)
